- Add vLLM - Tweak chub ui
