+++ web/pages/Chub/Chub.tsx
@@ -11,11 +11,13 @@ import ChubImportCharModal from './ChubImportChar'
 import ChubImportBookModal from './ChubImportBook'
 import { useNavigate, useParams } from '@solidjs/router'
 import { Page } from '/web/Layout'
+import { ListFilter } from 'lucide-solid'
+import Button from '/web/shared/Button'
+import Modal from '/web/shared/Modal'
 
 const chubTabs = {
   characters: 'Characters',
   lorebooks: 'Lorebooks',
-  filter: 'Filter Settings',
 }
 
 type Tab = keyof typeof chubTabs
@@ -23,12 +25,13 @@ type Tab = keyof typeof chubTabs
 const Chub: Component = () => {
   const params = useParams()
   const nav = useNavigate()
-  const tabs: Tab[] = ['characters', 'lorebooks', 'filter']
+  const tabs: Tab[] = ['characters', 'lorebooks']
 
   const [tab, setTab] = createSignal(0)
   const [charModal, setCharModal] = createSignal<{ char: NewCharacter; fullPath: string }>()
   const [bookModal, setBookModal] = createSignal<{ book: AppSchema.MemoryBook; fullPath: string }>()
   const [loading, setLoading] = createSignal(false)
+  const [filter, setFilter] = createSignal(false)
 
   createEffect(() => {
     const index = tabs.indexOf(params.tab || ('characters' as any))
@@ -44,7 +47,11 @@ const Chub: Component = () => {
     <Page>
       <PageHeader title="Character Hub" />
 
-      <div class="my-2">
+      <div class="my-2 flex gap-2">
+        <Button schema="clear" onClick={() => setFilter(true)}>
+          <ListFilter />
+        </Button>
+
         <Tabs
           tabs={tabs.map((t) => chubTabs[t])}
           selected={tab}
@@ -66,10 +73,6 @@ const Chub: Component = () => {
 
         <div class={currentTab() === 'lorebooks' ? tabClass : 'hidden'}>
           <BookList setBook={(book, fullPath) => setBookModal({ book, fullPath })} />
-        </div>
-
-        <div class={currentTab() === 'filter' ? tabClass : 'hidden'}>
-          <FilterSettings />
         </div>
       </div>
       <div>
@@ -94,6 +97,13 @@ const Chub: Component = () => {
           book={bookModal()!.book}
         />
       </Show>
+      <Modal
+        show={filter()}
+        close={() => setFilter(false)}
+        footer={<Button onClick={() => setFilter(false)}>Close</Button>}
+      >
+        <FilterSettings />
+      </Modal>
     </Page>
   )
 }
