+++ web/shared/GenerationSettings.tsx
@@ -142,6 +142,7 @@ const GenerationSettings: Component<Props & { onSave: () => void }> = (props) =>
               { label: 'Textgen (Ooba)', value: 'ooba' },
               { label: 'Llama.cpp', value: 'llamacpp' },
               { label: 'Ollama', value: 'ollama' },
+              { label: 'vLLM', value: 'vllm' },
               { label: 'Aphrodite', value: 'aphrodite' },
               { label: 'ExLlamaV2', value: 'exllamav2' },
               { label: 'KoboldCpp', value: 'koboldcpp' },
